requirement.txt
channels
cmake
cvlib
Django
django-browser-reload
django-cors-headers
djangorestframework
dlib
face-recognition
face-recognition-models
mysqlclient
numpy
opencv-python
pandas
pillow
requests
tensorboard
tensorflow